+++ project/src/main/java/com/company/banking/domain/bankaccount/PersonalAccount.java
@@ -1,7 +1,7 @@
 package com.company.banking.domain.bankaccount;
 
 import com.company.banking.domain.AccountEntry;
-import com.company.banking.domain.AccountType;
+import com.company.common.AccountType;
 import com.company.banking.observer.Observer;
 
 public class PersonalAccount extends BankAccount{
