+++ project/src/main/java/com/company/banking/strategy/BankingStrategy.java
@@ -1,6 +1,6 @@
 package com.company.banking.strategy;
 
-import com.company.base.Strategy;
+import com.company.common.Strategy;
 
 public interface BankingStrategy extends Strategy {
     public double calculateInterest(double balance);
