+++ project/src/main/java/com/company/common/AccountStrategyType.java
@@ -1,4 +1,4 @@
-package com.company.base;
+package com.company.common;
 
 public enum AccountStrategyType {
 SAVINGS, CHECKING
