+++ project/src/main/java/com/company/common/AccountType.java
@@ -1,4 +1,4 @@
-package com.company.banking.domain;
+package com.company.common;
 
 public enum AccountType {
     CREDIT,
