+++ project/src/main/java/com/company/creditcard/CreditCardStrategy/CreditCardStrategy.java
@@ -1,6 +1,6 @@
 package com.company.creditcard.CreditCardStrategy;
 
-import com.company.base.Strategy;
+import com.company.common.Strategy;
 
 public interface CreditCardStrategy extends Strategy {
 }
